waiter: BREAKING DB: make isOwner nullable
The way the unique constraint was set up meant that, instead of there being only one owner, there could be only one owner and only one non-owner.

This is fixed now, by making it so, in the DB, the value of isOwner is NULL if you are not an owner and 1 or true if you are an owner.
This needs to be accounted for in the future — i.e. when taking away ownership, it should set it to NULL, not to false.